[unnecessary copy] Suppress issues on trivially copyable
Summary:
This diff suppresses issues on trivially copyable.

Problem: The unnecessary copy checker had checked if the target type is trivially copyable, but it
was not good enough, because sometimes the source is trivially copyable, but the target is not. For
example,

```
// std::string_view v;
std::optional<std::string> s = v;
```

it does not make sense to suggest `std::move(v)` here, while the developer may need to think about
if the new string construction is avoidable in high-level.

To avoid the false positive, this diff checks if the source type is trivially copyable or not,
together.

Reviewed By: jvillard

Differential Revision: D50796512

fbshipit-source-id: f20836a627c484ca6e2f4b970359dc6b012405db