[kotlin] Add FN test for null argument passed to third-party kotlin function
Summary:
Add FN test for null argument passed to third-party kotlin function

Before this change in order to check if proc is a Kotlin proc we were relying on the fact that we have access to source code which doesn't work for third party libraries. This diff illustrates that even without doing that we still don't handle procs coming from third party code but for a different reason now.

Reviewed By: dulmarod

Differential Revision:
D50835027

Privacy Context Container: L1122176

fbshipit-source-id: e53a6d3e782c4fca6d87916daa1e89b163c657d4