[hack] Make Hack driver respect capture-block-list and skip-analysis-in-path
Summary:
The exclusion is similar to what clang does. One downside is that `skip-analysis-in-path` also skips
capture. We may want to reconsider it in future in case we need to capture a file to fill in the
gaps in the type hierarchy but don't analyze it because it's too expensive.

Reviewed By: geralt-encore

Differential Revision: D50881405

fbshipit-source-id: 0067b697a9eb1b1e40e1d3d0e4d3e0fbd241e586