Map support in Lineage
Summary:
The preexisting map shape information allows Lineage to split some graph cells that represent a map
into different cells for each atom key. However, without explicit support in Lineage for the map
functions, most uses of maps were still imprecise because setting or getting an element would be
considered as an unknown procedure, therefore generating flow from-everything-to-everything.

We add specific models for the map-related functions that make use of the shape informations. When
used with a key that can be statically inferred as a constant set of atoms, these functions will only
add flow from or to the corresponding sub-cells of the map. When the key does not verify that
property, the models fall back to a model that writes to or reads from every cell of the map.

Reviewed By: rgrig

Differential Revision: D49159601

fbshipit-source-id: 76eb159b470216b2ad09cca94dd7541203f53dbe