[pulse][interproc] correctly handle restricted variables in preconditions
Summary:
With D49640217 we normalise in preconditions, which means now some values in
the precondition may be "restricted", i.e. belong to a special class of
`AbstractValue.t` that implies they denote a non-negative value. However
D49640217 didn't account for that, only for potential equalities between
abstract values in the precondition. As a result:
- the implicit assumption carried by the presence of a restricted abstract
  value in the pre isn't taken into account, leading to possibly some latent
  issues becoming manifest by mistake
- the non-negativity of these values can be lost when applying summaries

This fixes both issues.

However, neither issue is actually a problem at the moment because we still get
a trivially true `a >= 0` fact in the path condition (with `a` restricted,
hence `a >= 0` by construction) that gets propagated correctly due to
weaknesses in the arithmetic. The next diff will recognise and simplify such
trivially-true atoms so this change is needed first.

Reviewed By: skcho

Differential Revision: D50556737

fbshipit-source-id: c34ca63b9bcc142826c9e6dad5359856b1824eab