[pulse][formula] incremental term eq normalisation
Summary:
Keep `term_eqs` up to date wrt new equalities. Similarly to `linear_eqs` this
is done by keeping track of which variable appears in which term for
(relatively) cheap updates when said variables get equated to other variables
or to constants (these are the only cases needed as terms in `term_eqs` are
shallow).

Reviewed By: skcho

Differential Revision: D50556729

fbshipit-source-id: 242fc82f16724d6eb8af23bfc91ddf9d40ca29ab