[pulse][formula] incremental intervals normalisation
Summary:
Like the other domains in previous diffs, try to maintain a normalised form for
intervals so we don't need to normalise them. Unlike other domains this isn't
wired into the `Normalizer` module but instead we use the computed `new_eqs`
and incorporate them into intervals. This is because intervals already live
outside of `Normalizer` (and should be equivalent).

Reviewed By: skcho

Differential Revision: D50556728

fbshipit-source-id: 5f7b84055e2eb2cca58ebbc9b8afc67855802c3d