[pulse] stop exploring summaries past invalid addresses
Summary:
When we apply a summary we begin by fully exploring its precondition and
materialising it into the caller's current state. Then we check if any of the
addresses needed by the callee in this precondition were invalid in the caller.
This is not great for two reasons: 1) it is wonky, we shouldn't dereference
invalid addresses in the caller, 2) it interacts badly with the contradiction
detection that is triggered when a address that appears allocated is equal to
zero.

For example (let's pretend all values have been renamed to match in the caller and callee):
```
current state: x|->x' * x'=0 * Invalid(x')
callee precondition:  x|->x' * x'|->x'' * MustBeValid(x')
resulting caller state:
x|->x' * x'|->x'' * x'=0 * Invalid(x')
```

This state is equivalent to false because of `x'|->- * x'=0` but really what we
wanted to do here was report an error because `x'` is invalid in the caller but
dereferenced by the callee.

This diff stops the exploration from addresses that are invalid. Crucially, we
don't stop the overall exploration though. This is to 1) discover other
potential such errors, maybe one of them happens first in the callee and should
be reported instead of the first one we found, and 2) discover potential
contradictions that could mean this summary was inapplicable and we shouldn't
report.

Example, same as before but the caller has `y'=42` and the callee `y'≠42`
```
current state: x|->x' * x'=0 * Invalid(x') * y|->y' * y'=42
callee precondition:  x|->x' * x'|->x'' * MustBeValid(x') * y|->y' * y'≠42
resulting caller state: UNSAT
```

If we stop too soon we can miss that `y'` in the caller matches `y'` in the
callee and so that the formulas contradict each other.

Reviewed By: skcho

Differential Revision: D50931122

fbshipit-source-id: 71d3c5e5653cf13e1a1d55ac3dcb52d914a59455