[pulse] normalise caller addresses on the fly when applying summaries
Summary:
Exploring the precondition, and especially the formula of a summary can make us
learn new variable equalities that invalidate what was stored in (the range of)
the "callee abstract val -> caller abstract val" map. We didn't tackle this in
the earlier changes for on-the-fly canonicalisation, time to do so! This also
lets us ingest the entire callee formula in one go and avoids eager
normalisation of the substitution map.

Fixes https://github.com/facebook/infer/issues/1785

Reviewed By: skcho

Differential Revision: D50931121

fbshipit-source-id: 9c761e9f39720c022ef02e758dbc439cac211d5b