[infer] Sanitize issue qualifier more thoroughly to avoid spurious introduced errors
Summary:
Issue qualifier is a part of issue hash that we use during report
diffing. The sanitization that we used before didn't properly sanitize
qualifiers for taint error messages, which led to detecting introduced
errors on seemingly innocent changes. One such example is when a lambda
captures another variable, it shift the argument numbers for the taint
sink from e.g. `argument #1` to `argument #2` which in turn affects the
issue's hash.

This diff adds 'argument #N' pattern to the regex and makes the whole
thing case insensitive.

Reviewed By: davidpichardie

Differential Revision: D50990931

fbshipit-source-id: 044d70487e0f6f84954f80a563fc418499a057be