[uninit][hack] Run dynamic dispatch in hack_get_class
Summary:
This diff runs the dynamic dispatch in `hack_get_class`.

Problem: The `hack_get_class` is introduced when `new static()` is invoked. For example, it is
translated to something like,

```
__factory(hack_get_class($this));
```

However, `hack_get_class`' semantic model was the identity function, therefore it was able to pass
the *non-dynamic* type of `$this` to the `__factory`, which is infeasible, then this introduced false
positives.

Reviewed By: davidpichardie

Differential Revision: D51158645

fbshipit-source-id: e847f366f9bca7a9415178879172f913609c77b4