Include metadata in Lineage hash-based deduplication
Summary:
On one hand, Lineage uses deduplication based on hashing the edge components to avoid generating too
much redundant stuff. On the other hand, some edges that are distinguishable in the internal
representation may have the same "components" in after exporting: this is notably the case for Copy
edges from the same variable into fields of the return. Those fields are distinct nodes in the
internal graph, but are exported as the single `$ret` node on the output. Therefore those Copy edges
from the same source into different fields are exported as edges from the same source into the same
source, the field information being exported as metadata.

This diff includes this metadata in the hashing of edges, to make sure those edges are not
deduplicated away (meaning only one field will have its edge in the final graph).

Reviewed By: rgrig

Differential Revision: D51082662

fbshipit-source-id: 4dcc9ef9f6027b4371aaee051596d7b5734ad895