[pulse] fix missing cell id summary mapping for aliased locations
Summary:
When applying a summary precondition, we remember each visited address and stop
the exploration when we encounter an already-visited address. So far so good,
but in case of such aliasing we should still record the history that
corresponds to the cell id, which is different for different aliases to the
same value!

Reviewed By: skcho

Differential Revision: D51616240

fbshipit-source-id: a12537f9bb999f45f00a6c7449c9d589fc933c56