[reference-stability] Add sequenced-before `operator=` FP example
Summary:
Add a demo test case for this issue. It affects reference-stability but maybe other checkers now or in the future.

In C++, for an assignment operator, the RHS is guaranteed to be sequenced before the LHS. So here, `map[71]` happens first, then `/ 31` (no more dangling references), then `map[13]`, then the final assignment.

SIL currently computes the left hand side first.

Reviewed By: skcho

Differential Revision: D51729738

fbshipit-source-id: d2b96d7d3330dedd57f2913b01715add1ea87c00