Lineage: make PartialGraph type abstract
Summary:
As another step to improving the graph implementation, we decouple the final graph result with the
partial one accumulated on each procedure node by making the latter an abstract type.

The implementation itself doesn't change, only some functions are moved into the `PartialGraph`
module. It is actually likely that `PartialGraph` will remain as a list of edges as this is probably
the fastest way to accumulate edges until they can be consolidated altogether into the final graph.

Reviewed By: rgrig

Differential Revision: D51622863

fbshipit-source-id: 5a4b44409ee9d906989d95ca38a6cb562ab4f052