[pulse] more flexible way to propagate taint dependencies
Summary:
Instead of building the taint dependency graph then calling some function on it
iteratively we can do both at once. This saves the intermediate graph and lets
us tweak when we call the callback and how we keep exploring based on the
result of the callback if needed (whereas before the exploration is done
without taking the callback into consideration).

Reviewed By: skcho

Differential Revision: D52082315

fbshipit-source-id: d952249e7955b659eda29c3641390f1327480a7d