[pulse] use history for unknown calls, attributes for config & models
Summary:
For unknown calls, the value history should contain the taint event being
propagated as this is handled by value histories natively. For other cases it's
not the case and Hard to reconciliate how they work with how histories work at
the moment (need to back propagate histories to where the history was built
from, I think).

Anyhow this fixes some annoying FPs.

Reviewed By: rgrig

Differential Revision: D52082312

fbshipit-source-id: d6476e2001aa6d6774c6af0e40733acc76e6bb62